+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- **Added Icon Support:** Buttons can now display icons next to text using `primeButtonIcon` and `secondaryButtonIcon` properties.
+- **Added Button Shape Customization:** Introduced `buttonShape` property to allow custom `OutlinedBorder` for buttons (e.g., `RoundedRectangleBorder`).
+- **Added Gradient Background Support:** Buttons can now have gradient backgrounds using `primeButtonGradient` and `secondaryButtonGradient` properties.
+
 ## 0.0.1
 
 - initial release.